TASK41: Moved Types into own package folder and adjusted Imports
