+++ classes/__init__.py
@@ -1,0 +1,3 @@
+from .base_classes import *
+
+from .rectangle_packer_types import *
